+++ tests/data/cases/type_ignore_wrapped_string_in_dict.py
@@ -1,0 +1,15 @@
+# flags: --unstable
+my_dict = {
+    "key": (
+        "A very very very very very very very very very very very very very very very very long string literal"
+    )  # type: ignore
+}
+
+# output
+
+my_dict = {
+    "key": (
+        "A very very very very very very very very very very very very very very very"
+        " very long string literal"
+    )  # type: ignore
+}